+++ CHANGELOG.md
@@ -125,6 +125,7 @@
 - PIM-10094: Increase product grid filters limit display in user settings
 - PIM-10101: Fix incorrect count of families selection in the family grid
 - PIM-10107: Hide add quantified association button when ACL is not granted
+- PIM-10100: Added a product & product model indexation step for setAttributeRequirements job
 - PIM-10092: Display an error message when trying to delete a category tree linked to a channel
 
 ## New features